// apps/frontend/src/app/core/api/models/integrity-link-rule.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RuleType } from '../models/rule-type';
import { RuleValue } from '../models/rule-value';
export interface IntegrityLinkRule {
  group_or_role: string;
  id?: (number | null);
  integrity_link_id?: (string | null);
  rule_type: RuleType;
  rule_value?: RuleValue;
}